put the resolveAvailableProject into filterProjects
